Proofread and clarify documentation
If I was uncertain what a parameter means (for example scoring_matrix in 
customParams) I left it as is even if there was something to correct 
grammatically. I noticed that many functions didn't have a separate 
description or title, so I clarified the documentation a bit by 
splitting out a shorter title and a more elaborate description. I didn't 
do this if the text was very short, in which case there is a shared 
title and description. I barely 
wrote anything new in the process, so there shouldn't be any cases of me 
misunderstanding functionality.